Much more logic and clear way to use the lib
Clearly separate concrete grid and abstract stroke, without mixing interdistance of physical actuators and the stroke for example. Now, you compute stroke parameters GIVEN a specific grid, but the stroke exists outside the grid and can be used by different grids. Before the two were so coupled that it does not make sense that they are not the same Class